Move EstimatePicker to the components folder.
